隔离我自己用的 coffers 签到系统，提供了只阅读文章和看视频的入口
   coffers
   是我自己建的一个网站，我通过向其签到，请求其在未签到时提醒我学习强国。
   因为 coffers 网站没有多用户功能，所以默认关闭该功能，如果需要打开，
   请在 coffers.cfg 中配置自己的信息，并在 coffers.py 中更改代码逻辑。

+++ xuexi.py
@@ -17,51 +17,56 @@ p.wait()
 
 while True:
 
-    print("=============引导菜单==============")
+    print("=====================引导菜单====================")
     print("      0. 全部得分任务")
     print("      1. 全部得分任务与足额点点通")
-    print("      2. 阅读文章时间得分任务")
-    print("      3. 阅读文章次数得分任务")
-    print("      4. 观看视频时间得分任务")
-    print("      5. 观看视频次数得分任务")
-    print("      6. 挑战答题得分任务")
+    print("      2. 阅读文章与观看视频得分任务与足额点点通")
+    print("      3. 阅读文章时间得分任务")
+    print("      4. 阅读文章次数得分任务")
+    print("      5. 观看视频时间得分任务")
+    print("      6. 观看视频次数得分任务")
+    print("      7. 挑战答题得分任务")
     print("      x. 退出程序")
-    print("===================================")
+    print("=================================================")
     p = say("请输入数字选择引导得分任务")
     sel = input()
     p.kill()
 
     if sel != '0' and sel != '1' and sel != '2' and sel != '3' and sel != '4' \
-                  and sel != '5' and sel != '6' and sel != 'x':
+                  and sel != '5' and sel != '6' and sel != '7' and sel != 'x':
         p = say("输入有误，请重新输入数字选择引导模块，模块编号请参考屏幕输出")
         sel = input()
         p.kill()
 
-    if sel == '2':
-        p_read_article(6, 60)
     if sel == '3':
-        p_read_article(6, 5)
+        p_read_article(6, 60)
     if sel == '4':
-        p_watch_video(6, 60)
+        p_read_article(6, 5)
     if sel == '5':
-        p_watch_video(6, 5)
+        p_watch_video(6, 60)
     if sel == '6':
+        p_watch_video(6, 5)
+    if sel == '7':
         p_challenge_answer()
-    if sel == '0' or sel == '1':
+    if sel == '0' or sel == '1' or sel == '2':
         time_start = time.time()
 
-        p_daily_answer()
-        p_weekend_answer()
-        p_specific_answer()
-        print("提示：点点通任务需要答对 15 道挑战答题题目")
-        p_challenge_answer()
-        p_likesome()
-        p_watch_local_video()
+        if sel != '2':
+            p_daily_answer()
+            p_weekend_answer()
+            p_specific_answer()
+            print("提示：点点通任务需要答对 15 道挑战答题题目")
+            p_challenge_answer()
+            p_likesome()
+            p_watch_local_video()
+
         p_read_article(6, 60)  # 阅读 6 篇，每篇 60 秒
         p_watch_video(6, 60)  # 收听 6 个，每个 60 秒
-        p_fightforup_answer()
 
-        if sel == '1':
+        if sel != '2':
+            p_fightforup_answer()
+
+        if sel != '0':
             print("提示：点点通任务需要阅读 12 篇文章、观看 12 个视频")
             p_read_article(6, 5)  # 阅读 6 篇，每篇 5 秒
             p_watch_video(6, 5)  # 收听 6 个，每个 5 秒
@@ -71,16 +76,19 @@ while True:
         p = say("总用时" + str(int(time_during)) + "分钟")
         p.wait()
 
-    if sel == 'x' or sel == '0' or sel == '1':
+    if sel == 'x' or sel == '0' or sel == '1' or sel == '2':
         # qiandao to coffers
-        site = "http://ylm.pythonanywhere.com"
-        ctc = Connect_to_coffers(site)
-        r = ctc.qiandao(True)
-        if r == 200:
-            p = say("已通知 coffers 标记完成任务")
+        ctc = Connect_to_coffers()
+        if not ctc.is_open():
+            p = say("coffers 未启用")
+            p.wait()
         else:
-            p = say("未联系到 coffers，标记失败")
-        p.wait()
+            r = ctc.qiandao(True)
+            if r == 200:
+                p = say("已通知 coffers 标记完成任务")
+            else:
+                p = say("未联系到 coffers，标记失败")
+            p.wait()
 
         say("程序正在退出，祝您生活愉快")
         p.wait()
